Support microtasks, rendering, and browser tasks
This commit expands the definition of what a long task is. Before, a long task was equivalent to an event loop task that runs for a long time. With this change, a long task can come from an event loop task plus associated microtasks, a rendering update step, or browser work outside of the event loop.

Attribution for the new types of long tasks is skipped but the type of task is set in the |name| of the  PerformanceLongTaskTiming entry. The name is a bit overloaded because it already contains the frame attribution, so if needed a new |taskType| attribute can be added, but then it's unclear what |name| would be for the new long task types.

This commit also fixes some problems found along the way. Most notably, everything in the reporting algorithm following "For each destinationRealm in destinationRealms" must be a substep of this for loop because the last step consists of queueing the entry in the corresponding |destinationRealm|.