Remove "browser" and "rendering" task types
From TPAC discussions, we decided to not provide "rendering" and "browser" types for now. The task attribution improvement is delayed until we have a good idea of what types of tasks we should be exposing which developers would find useful. The fear here is that exposing types incrementally could be confusing to developers. This change "fixes" https://github.com/w3c/longtasks/issues/40

No test changes required because I had not updated tests for this. Will followup by changing the caller of the Report long task algorithm in HTML spec (removing the |type| parameter).