Fix containerType for tasks without culprit
Currently, the longtasks spec uses 'Culprit browsing context container' to identify the browsing context container that caused a long task, when there is a single one that can be pointed at. This PR fixes containerType by setting it to 'window' when no single iframe is found to be the culprit of a longtask. It also replaces some nulls with empty strings because the attributes are not nullable.
Fixes https://github.com/w3c/longtasks/issues/61