Refactor report long task
This PR changes the algorithm so that it reports a long task for a single Document each time it is called. It also makes the task attribute mandatory as we do not intend to report long tasks from rendering. Will fix https://github.com/w3c/longtasks/issues/75 and https://github.com/w3c/longtasks/issues/80 when merged alongside the HTML changes.